feat: add ispc
ispc is a compiler for a variant of the C programming language, with
extensions for single program, multiple data programming. Under the
SPMD model, the programmer writes a program that generally appears
to be a regular serial program, though the execution model is actually
that a number of program instances execute in parallel on the hardware.

Signed-off-by: WangYuli <wangyl5933@chinaunicom.cn>